moved main.go to docs/ instead of docs/src/ because it looked strange not having an entrypoint in the docs/ folder. now it's more obvious where the program starts and what to do. also the package.json has a npm run build and npm run tw-build. also added tailwindcss/typography to package.json
